add file
Given the following strings: A = "The early bird catches the worm" B = "Patience is a virtue" Your task is to extract the word "early" from A and "virtue" from B. Then, concatenate these two words to form a sentence. After that, capitalize the sentence and find the last occurrence of the letter 'V' from the capitalized sentence. Perform all of these tasks using proper String class functions.